Rename `PushHandlingWakeLock` to `FetchPushForegroundServiceManager`.
Move the start/stop logic from `FetchPushForegroundService.Companion` to it.
